feat(component): 新增子分组列表tab
